linza-board/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
sqlalchemy==2.0.36
alembic>=1.13
psycopg2-binary>=2.9
boto3==1.35.85
httpx>=0.27.0
python-json-logger>=3.0
slowapi>=0.1.9
